requirements.txt
pyyaml
jsonschema
click
colorama
